refactor to HostDisplay
