+++ jparty/game.py
@@ -2,15 +2,7 @@ from PyQt6.QtCore import Qt, QObject, pyqtSignal
 from PyQt6.QtWidgets import (
     QInputDialog,
     QApplication,
-    QDialog,
-    QVBoxLayout,
-    QPushButton,
-    QSpinBox,
-    QLabel,
-    QGraphicsDropShadowEffect
     )
-
-from PyQt6.QtGui import QColor
 
 import threading
 import time
@@ -384,18 +376,6 @@ class Game(QObject):
         self.current_round = self.data.rounds[i - 1]
         self.dc.board_widget.load_round(self.current_round)
 
-    def set_player_in_control(self, new_player):
-        for player in self.players:
-            # Remove glow around player widget
-            self.host_display.player_widget(player).setGraphicsEffect(None)
-
-        if new_player is not None:
-            # Add white glow around player widget with offset 0, 0
-            effect = QGraphicsDropShadowEffect()
-            effect.setColor(QColor(255, 255, 255, 255))
-            effect.setBlurRadius(100)
-            effect.setOffset(0, 0)
-            self.host_display.player_widget(new_player).setGraphicsEffect(effect)
 
     def next_round(self):
         logging.info("next round")
@@ -407,14 +387,14 @@ class Game(QObject):
 
         self.current_round = self.data.rounds[i + 1]
         if isinstance(self.current_round, FinalBoard):
-            self.set_player_in_control(None)
+            self.host_display.set_player_in_control(None)
 
             self.dc.load_final(self.current_round.question)
             self.start_final()
         else:
             # Highlight player with least money to have control
             losing_player = min(self.players, key=lambda p: p.score)
-            self.set_player_in_control(losing_player)
+            self.host_display.set_player_in_control(losing_player)
 
             self.dc.board_widget.load_round(self.current_round)
 
@@ -577,7 +557,7 @@ class Game(QObject):
             self.answering_player,
             self.answering_player.score + self.active_question.value,
         )
-        self.set_player_in_control(self.answering_player)
+        self.host_display.set_player_in_control(self.answering_player)
         self.dc.borders.lights(False)
 
         if self.active_question.dd:
